+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/advancedsettings/view/AdvancedSettingsView.kt
@@ -29,7 +29,6 @@ import co.electriccoin.zcash.ui.design.component.ZashiButton
 import co.electriccoin.zcash.ui.design.component.ZashiButtonDefaults
 import co.electriccoin.zcash.ui.design.component.ZashiHorizontalDivider
 import co.electriccoin.zcash.ui.design.component.ZashiSettingsListItem
-import co.electriccoin.zcash.ui.design.component.ZashiSettingsListItemState
 import co.electriccoin.zcash.ui.design.component.ZashiSmallTopAppBar
 import co.electriccoin.zcash.ui.design.component.ZashiTopAppBarBackNavigation
 import co.electriccoin.zcash.ui.design.newcomponent.PreviewScreens
@@ -37,7 +36,6 @@ import co.electriccoin.zcash.ui.design.theme.ZcashTheme
 import co.electriccoin.zcash.ui.design.theme.colors.ZashiColors
 import co.electriccoin.zcash.ui.design.theme.dimensions.ZashiDimensions
 import co.electriccoin.zcash.ui.design.util.orDark
-import co.electriccoin.zcash.ui.design.util.stringRes
 import co.electriccoin.zcash.ui.screen.advancedsettings.AdvancedSettingsState
 import co.electriccoin.zcash.ui.screen.advancedsettings.AdvancedSettingsTag
 
@@ -71,13 +69,13 @@ fun AdvancedSettings(
         ) {
             ZashiSettingsListItem(
                 text = stringResource(id = R.string.advanced_settings_recovery),
-                icon = R.drawable.ic_advanced_settings_recovery orDark R.drawable.ic_advanced_settings_recovery_dark,
+                icon = R.drawable.ic_advanced_settings_recovery,
                 onClick = state.onRecoveryPhraseClick
             )
             ZashiHorizontalDivider()
             ZashiSettingsListItem(
                 text = stringResource(id = R.string.advanced_settings_export),
-                icon = R.drawable.ic_advanced_settings_export orDark R.drawable.ic_advanced_settings_export_dark,
+                icon = R.drawable.ic_advanced_settings_export,
                 onClick = state.onExportPrivateDataClick
             )
             ZashiHorizontalDivider()
@@ -85,7 +83,7 @@ fun AdvancedSettings(
                 text = stringResource(id = R.string.advanced_settings_choose_server),
                 icon =
                     R.drawable.ic_advanced_settings_choose_server orDark
-                        R.drawable.ic_advanced_settings_choose_server_dark,
+                        R.drawable.ic_advanced_settings_choose_server,
                 onClick = state.onChooseServerClick
             )
             ZashiHorizontalDivider()
@@ -93,16 +91,9 @@ fun AdvancedSettings(
                 text = stringResource(id = R.string.advanced_settings_currency_conversion),
                 icon =
                     R.drawable.ic_advanced_settings_currency_conversion orDark
-                        R.drawable.ic_advanced_settings_currency_conversion_dark,
+                        R.drawable.ic_advanced_settings_currency_conversion,
                 onClick = state.onCurrencyConversionClick
             )
-            if (state.coinbaseButton != null) {
-                ZashiHorizontalDivider()
-                ZashiSettingsListItem(
-                    icon = R.drawable.ic_advanced_settings_coinbase,
-                    state = state.coinbaseButton
-                )
-            }
             Spacer(modifier = Modifier.height(24.dp))
             Spacer(modifier = Modifier.weight(1f))
             Row(
@@ -171,11 +162,6 @@ private fun AdvancedSettingsPreview() =
                     onChooseServerClick = {},
                     onCurrencyConversionClick = {},
                     onDeleteZashiClick = {},
-                    coinbaseButton =
-                        ZashiSettingsListItemState(
-                            text = stringRes("Coinbase"),
-                            onClick = {}
-                        )
                 ),
             topAppBarSubTitleState = TopAppBarSubTitleState.None,
         )